perf: convert /titles and /laws list pages to SSR with ISR (closes #269)
Both pages are now async Server Components with revalidate=300 (5 min ISR).
Data arrives with the HTML — no client-side waterfall on initial load.

Sorting on /laws stays client-side in the new LawsTable component, which
receives the pre-fetched law list as props from the server component.

Server-side fetches use BACKEND_URL directly (lib/api.server.ts) instead
of the /api rewrite path, which only applies to client-originated requests.

Updated page.test.tsx to render `await TitlesPage()` (async server
component pattern) and removed the now-unnecessary QueryClientProvider.
