+++ frontend/src/app/titles/page.test.tsx
@@ -1,6 +1,5 @@
-import { render, screen, waitFor } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import TitlesPage from './page';
 
 vi.mock('next/link', () => ({
@@ -18,6 +17,10 @@ vi.mock('next/link', () => ({
   ),
 }));
 
+vi.mock('next/navigation', () => ({
+  usePathname: () => '/titles',
+}));
+
 const titles = [
   {
     title_number: 10,
@@ -28,13 +31,6 @@ const titles = [
     section_count: 200,
   },
 ];
-
-function wrapper({ children }: { children: React.ReactNode }) {
-  const qc = new QueryClient({
-    defaultOptions: { queries: { retry: false } },
-  });
-  return <QueryClientProvider client={qc}>{children}</QueryClientProvider>;
-}
 
 describe('TitlesPage', () => {
   beforeEach(() => {
@@ -50,20 +46,16 @@ describe('TitlesPage', () => {
     vi.restoreAllMocks();
   });
 
-  it('renders the page heading after loading', async () => {
-    render(<TitlesPage />, { wrapper });
-    await waitFor(() => {
-      expect(screen.getByRole('heading', { level: 1 })).toHaveTextContent(
-        'Browse the US Code'
-      );
-    });
+  it('renders the page heading', async () => {
+    render(await TitlesPage());
+    expect(screen.getByRole('heading', { level: 1 })).toHaveTextContent(
+      'Browse the US Code'
+    );
   });
 
-  it('renders titles as directory items after loading', async () => {
-    render(<TitlesPage />, { wrapper });
-    await waitFor(() => {
-      expect(screen.getByText(/Armed Forces/)).toBeInTheDocument();
-    });
+  it('renders titles as directory items', async () => {
+    render(await TitlesPage());
+    expect(screen.getByText(/Armed Forces/)).toBeInTheDocument();
     expect(screen.getByRole('link', { name: /Armed Forces/ })).toHaveAttribute(
       'href',
       '/titles/10'
